+++ dev/user.clj
@@ -3,7 +3,8 @@
             [clojure.walk :refer [postwalk]]
             [parcera.core :as parcera]
             [grape.core :as g]
-            [grape.cli :as cli]))
+            [grape.cli :as cli]
+            [grape.impl.models :as m]))
 
 (def s
   (slurp "src/grape/core.clj"))