+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Grape Changelog
 
 ## Unreleased
-
+### CLI
 * Add `-v`/`--version` to the command-line
 * Add `-c`/`--count` to show the total matches count
 * Remove `-r`/`--recursive`: it’s always recursive now
+
+### Internals
+* Extract models code from `grape.core` into `grape.impl.models`
 
 ## 0.3.0 (2020/10/19)